# docs/schemas/openapi.yaml
openapi: 3.1.0
info:
  title: CherryGraph Studio API
  version: 0.5.0
  description: |
    CherryGraph Studio — 知识图谱驱动的 Wiki + RAG Chat 平台。
    所有写入类接口支持 X-Idempotency-Key header（24h 去重窗口）。
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    bridgeBearer:
      type: http
      scheme: bearer
      description: DOCMOST_BRIDGE_SECRET as Bearer token for internal Bridge endpoints
    bridgeHmacAuth:
      type: apiKey
      in: header
      name: X-Bridge-Signature
      description: HMAC-SHA256(payload, DOCMOST_BRIDGE_SECRET)
    workerApiKey:
      type: apiKey
      in: header
      name: X-Worker-Key
      description: WORKER_API_KEY for internal Worker-to-API endpoints
...
  schemas:
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        per_page: { type: integer }
        total: { type: integer }
        has_next: { type: boolean }
    Meta:
      type: object
      properties:
        request_id: { type: string }
        pagination: { $ref: '#/components/schemas/Pagination' }
    ErrorResponse:
      type: object
      required: [error, meta]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object }
        meta: { $ref: '#/components/schemas/Meta' }

    # --- Auth ---
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    BrowserLoginResponse:
      type: object
      required: [access_token, expires_in, user]
      properties:
        access_token: { type: string }
        expires_in: { type: integer }
        user:
          type: object
          required: [id, email, name, role, groups]
          properties:
            id: { type: string }
            email: { type: string, format: email }
            name: { type: string }
            role: { type: string, enum: [owner, admin, space_admin, editor, viewer, auditor] }
            groups:
              type: array
              items: { type: string }
    BrowserRefreshResponse:
      type: object
      required: [access_token, expires_in]
      properties:
        access_token: { type: string }
        expires_in: { type: integer }

    # --- User ---
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string }
        role: { type: string, enum: [owner, admin, space_admin, editor, viewer, auditor] }
        status: { type: string, enum: [active, invited, disabled] }
        groups:
          type: array
          items: { $ref: '#/components/schemas/GroupRef' }
        spaces:
          type: array
          items: { $ref: '#/components/schemas/SpaceRef' }
        last_login_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    GroupRef:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
    SpaceRef:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        role: { type: string }
...
    PasswordChangeRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password: { type: string }
        new_password: { type: string, minLength: 8 }
    Session:
      type: object
      properties:
        id: { type: string }
        ip: { type: string }
        user_agent: { type: string }
        created_at: { type: string, format: date-time }
        last_used_at: { type: string, format: date-time }
        is_current: { type: boolean }
...
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: Invalid request parameters
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
paths:
...
  /auth/login:
    post:
      summary: Login
      description: Browser login returns an access token in the JSON body and sets the refresh token as an HttpOnly `refresh_token` cookie.
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Login successful. The refresh token is set in the `Set-Cookie` header and is not included in the JSON body.
          headers:
            Set-Cookie:
              description: HttpOnly `refresh_token` cookie scoped to `/api/auth`.
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/BrowserLoginResponse' }
                  meta: { $ref: '#/components/schemas/Meta' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '429': { $ref: '#/components/responses/RateLimitError' }
...
  /auth/me:
    get:
      summary: Get current user
      tags: [Auth]
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/User' }
                  meta: { $ref: '#/components/schemas/Meta' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
  /auth/refresh:
    post:
      summary: Refresh token
      description: Browser refresh reads the HttpOnly `refresh_token` cookie, rotates it, and returns only a new access token in the JSON body. No JSON request body is required or used.
      tags: [Auth]
      security: []
      responses:
        '200':
          description: Token refreshed. The rotated refresh token is set in the `Set-Cookie` header and is not included in the JSON body.
          headers:
            Set-Cookie:
              description: Rotated HttpOnly `refresh_token` cookie scoped to `/api/auth`.
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/BrowserRefreshResponse' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
  /auth/password/change:
    post:
      summary: Change password
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PasswordChangeRequest' }
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      success: { type: boolean }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /auth/sessions:
    get:
      summary: List active sessions
      tags: [Auth]
      responses:
        '200':
          description: Active sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Session' } }
                  meta: { $ref: '#/components/schemas/Meta' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
  /auth/sessions/{session_id}:
    delete:
      summary: Revoke session
      tags: [Auth]
      parameters:
        - { name: session_id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Session revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      revoked: { type: boolean }
        '404': { $ref: '#/components/responses/NotFoundError' }

  # --- Space ---
...